+++ encodings/decimal-byte-parts/src/decimal_byte_parts/testing.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-//! Test-only helpers for building byte-parts arrays.
+//! Shared fixtures for decimal byte-parts tests.
 
 use vortex_array::VortexSessionExecute;
 use vortex_array::array_session;
@@ -13,18 +13,12 @@ use vortex_buffer::Buffer;
 use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
 
-use crate::DecimalByteParts;
-use crate::DecimalBytePartsArray;
-use crate::decimal_byte_parts::limbs::split_decimal;
+use super::DecimalBytePartsArray;
+use super::dbp_encode;
 
 /// Encode a canonical decimal array as byte parts, splitting wide values into lower parts.
 pub(crate) fn encode(decimal: &DecimalArray) -> VortexResult<DecimalBytePartsArray> {
-    let parts = split_decimal(decimal, &mut array_session().create_execution_ctx())?;
-    DecimalByteParts::try_new_with_lower_parts(
-        parts.msp,
-        parts.lower_parts,
-        decimal.decimal_dtype(),
-    )
+    dbp_encode(decimal, &mut array_session().create_execution_ctx())
 }
 
 /// An `i128`-backed decimal array, encoded as byte parts with one lower part.
@@ -48,6 +42,6 @@ pub(crate) fn i256_parts(values: Vec<i256>, validity: Validity) -> DecimalBytePa
 }
 
 /// Build an `i256` from a signed high `i128` and unsigned low `u128`.
-pub(crate) fn i256_of(high: i128, low: u128) -> i256 {
+pub(super) fn i256_of(high: i128, low: u128) -> i256 {
     i256::from_parts(low, high)
 }
